HFDL: aircraft model lookup, extended info screen, decoder improvements, display fix
- Add ICAO24-based aircraft model lookup (HFDL_ICAO24_DB): reads a
  compact OpenSky-derived database (ICAO24.BIN) from the external SPI
  flash volume via a new read-only FAT12 file accessor. Includes a
  PC-side converter (scripts/make_icao24_compact.py) and a
  4KB-padded file layout so settings saves (which erase/rewrite their
  own 4KB flash block) can no longer corrupt the file header.

- Add HFDL aircraft table screen (HFDL_AIRCRAFT_TABLE): third view
  (spectrum -> reception log -> aircraft table), one row per aircraft
  with ICAO/ID, model, flight, position, time, TX side and event type,
  plus a header (decoder state, carrier offset, ground station for the
  tuned frequency via a generated systable) and footer (preamble
  funnel, CRC counters). No extra RAM: shares the existing text grid
  with the reception log.

- Fix: blank the 2px gap between the spectrum and waterfall panels on
  menu close, which previously left a stray strip of menu-tile colors
  under the spectrum after closing any menu (UI_PANEL_GAP_FIX, on by
  default).

- Add AIC3204_SINGLE_ENDED_TEST as a proper build-time option
  (Makefile-selectable) for bench testing with a line-level source
  direct to the codec, instead of the default differential/QSD wiring.

- Improve HFDL frame detection/decoding:
  - Normalized-LMS equalizer update (HFDL_EQ_NLMS, now default),
    matching liquid-dsp's reference behavior; the previous unnormalized
    update used an effective step ~15x too large, causing frequent
    training divergence.
  - HFDL_ABORT_BAD_TRAINING: give up on a frame early if equalizer
    training doesn't converge, freeing the receiver sooner for the
    next burst.
  - HFDL_CHAIN_ALWAYS_ON / HFDL_HOLD_MAX_DROPOUT_MS experimental knobs
    to work around the burst detector's poor sensitivity on
    RF-front-end noise (vs. the KiwiSDR audio path it was tuned for).
  - Added scripts/hfdl_log_stats.py and various debug-UART telemetry
    (input level, equalizer error, Costas offset, M1 correlation
    strength) to support A/B testing of the above.

All new behavior is opt-in via build flags (HFDL_ICAO24_DB,
HFDL_AIRCRAFT_TABLE, HFDL_EQ_NLMS=0/UI_PANEL_GAP_FIX=0 for legacy
behavior, etc.); default builds without these flags are unchanged
except where noted (equalizer default flip).

+++ User/hfdl_payload_decode.c
@@ -266,15 +266,24 @@ void hfdl_payload_decode_feed_symbol(float32_t eq_i, float32_t eq_q, uint32_t po
  * realistic use before someone checks the screen or reboots). last_ok
  * and last_valid packed into one byte (bit 0 = valid, bit 1 = ok) to
  * save the other byte a separate uint8_t would cost. Read via
- * hfdl_payload_decode_get_crc_status() below. */
-static uint8_t s_hfdl_crc_attempts = 0u;
-static uint8_t s_hfdl_crc_good = 0u;
+ * hfdl_payload_decode_get_crc_status() below.
+ *
+ * 18/09/2026: the two counters were widened to uint16_t. They are now
+ * shown permanently as "CRC OK good/attempts" on the HFDL reception
+ * log screen (see main.c's hfdl_log_panel_draw()), which is meant to be
+ * left running for hours on a busy channel - with a uint8_t pair,
+ * "attempts" wraps after 255 bursts (well under two hours at a few
+ * bursts per minute) and the two numbers then wrap independently of
+ * each other, giving nonsense such as "40/3". The cost is 2 bytes of
+ * RAM in total. */
+static uint16_t s_hfdl_crc_attempts = 0u;
+static uint16_t s_hfdl_crc_good = 0u;
 #define HFDL_CRC_STATUS_VALID_BIT 0x01u
 #define HFDL_CRC_STATUS_OK_BIT    0x02u
 static uint8_t s_hfdl_crc_status_bits = 0u; /* 0 until the first decode ever completes */
 
 void hfdl_payload_decode_get_crc_status(uint8_t *last_ok_out, uint8_t *last_valid_out,
-        uint8_t *attempts_out, uint8_t *good_out)
+        uint16_t *attempts_out, uint16_t *good_out)
 {
     if (last_ok_out) { *last_ok_out = (s_hfdl_crc_status_bits & HFDL_CRC_STATUS_OK_BIT) ? 1u : 0u; }
     if (last_valid_out) { *last_valid_out = (s_hfdl_crc_status_bits & HFDL_CRC_STATUS_VALID_BIT) ? 1u : 0u; }
